all api change done

+++ app/Http/Controllers/Api/V1/WishListController.php
@@ -23,7 +23,7 @@ class WishListController extends Controller
             ->leftJoin('vehicle_translations', 'wish_lists.vehicle_id', 'vehicle_translations.vehicle_id')
             ->where('vehicle_translations.locale', App::getLocale())
             ->where('wish_lists.user_id', $user->id)
-            ->select('wish_lists.id as wishlist_id','wish_lists.vehicle_id as wishlist_vehicle_id', 'vehicle_translations.name as vehicle_name', 'vehicle_translations.short_description', 'vehicle_translations.description', 'vehicles.*')
+            ->select('wish_lists.id as wishlist_id','wish_lists.vehicle_id as wishlist_vehicle_id', 'vehicle_translations.*', 'vehicles.*')
             ->get();
         $result = WishlistResource::collection($wishlist);
         return response()->json([
@@ -35,24 +35,33 @@ class WishListController extends Controller
     public function store(WishListStoreRequest $request)
     {
         $user = $request->user();
-        $count = DB::table('wish_lists')->where('vehicle_id', $request->vehicle_id)->where('user_id', $user->id)->count();
-        if ($count > 0) {
-            DB::table('wish_lists')->where('vehicle_id', $request->vehicle_id)->where('user_id', $user->id)->delete();
+        $vehicle_user_id =  DB::table('vehicles')->where('id', $request->vehicle_id)->first()->user_id;
+        if ($user->id != $vehicle_user_id){
+            $count = DB::table('wish_lists')->where('vehicle_id', $request->vehicle_id)->where('user_id', $user->id)->count();
+            if ($count > 0) {
+                DB::table('wish_lists')->where('vehicle_id', $request->vehicle_id)->where('user_id', $user->id)->delete();
+                return response()->json([
+                    'success' => true,
+                    'is_wishlist' => 0,
+                    'message' => trans('web_string.remove_in_wishlist_successfully'),
+                ]);
+            } else {
+                $wish_list = new WishList();
+                $wish_list->user_id = $user->id;
+                $wish_list->vehicle_id = $request->vehicle_id;
+                $wish_list->save();
+                return response()->json([
+                    'success' => true,
+                    'is_wishlist' => 1,
+                    'message' => trans('web_string.add_in_wishlist_successfully'),
+                ]);
+            }
+        }else{
             return response()->json([
-                'success' => true,
-                'is_wishlist' => 0,
-                'message' => trans('web_string.remove_in_wishlist_successfully'),
-            ]);
-        } else {
-            $wish_list = new WishList();
-            $wish_list->user_id = $user->id;
-            $wish_list->vehicle_id = $request->vehicle_id;
-            $wish_list->save();
-            return response()->json([
-                'success' => true,
-                'is_wishlist' => 1,
-                'message' => trans('web_string.add_in_wishlist_successfully'),
+                'success' => false,
+                'message' => trans('web_string.you_can_not_add_your_car_to_the_wish_list'),
             ]);
         }
+
     }
 }
